♻️  Work section in About me
