+++ src/components/pages/about/Experience/Job/CompanyLogo/index.tsx
@@ -1,14 +1,14 @@
-import styles from './styles.module.css'
+import Image from 'next/image'
 
 type Props = { url: string, name: string }
 
 const CompanyLogo = (props: Props) => (
-  <div className={styles.container}>
-    <img
+  <div>
+    <Image
+      className='rounded-xl hidden sm:block'
       alt={props.name}
-      className={`lazyload ${styles.logo}`}
-      data-src={props.url}
-      src='data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=='
+      blurDataURL='data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=='
+      src={props.url}
       width={48}
       height={48}
     />
