--- src/components/pages/about/Experience/Job/CompanyLogo/styles.module.css
@@ -1,10 +1,0 @@
-.container {
-  margin-right: 1em;
-}
-
-.logo {
-  border-radius: 50%;
-  min-height: 48px;
-  min-width: 48px;
-  max-width: 100%;
-}
