--- src/components/pages/about/ItemTimestamp/__tests__/stubs.ts
@@ -1,9 +1,0 @@
-export const props = {
-  dateStart: '12/15/2010',
-  dateFinish: '12/15/2015'
-}
-
-export const propsNoFinish = {
-  dateStart: '12/15/2010',
-  dateFinish: null
-}
