--- src/components/pages/about/ItemTimestamp/styles.module.css
@@ -1,7 +1,0 @@
-.timestamp {
-  color: var(--textSecondary);
-}
-
-.currentPosition {
-  font-weight: 700;
-}
