--- src/components/pages/about/ItemTitle/__tests__/stubs.tsx
@@ -1,7 +1,0 @@
-export const propsString = {
-  children: 'Title'
-}
-
-export const propsComponent = {
-  children: <h1>Component</h1>
-}
