--- src/components/pages/about/ItemTitle/styles.module.css
@@ -1,5 +1,0 @@
-.title {
-  margin: 0;
-  font-size: 1.15rem;
-  font-weight: 700;
-}
